fix: fix compiling warnings on windows

+++ src/rs_driver/driver/decoder/block_iterator.hpp
@@ -31,6 +31,7 @@ THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 *********************************************************************************************************************/
 
 #pragma once
+
 namespace robosense
 {
 namespace lidar
@@ -43,14 +44,14 @@ public:
 
   static const int MAX_BLOCKS_PER_PKT = 12;
 
-  void get(uint16_t blk, int32_t& az_diff, float& ts)
+  void get(uint16_t blk, int32_t& az_diff, double& ts)
   {
     az_diff = az_diffs[blk];
     ts = tss[blk];
   }
 
   BlockIterator(const T_Packet& pkt, uint16_t blocks_per_pkt, double block_duration, 
-      uint16_t block_az_duration, float fov_blind_duration)
+      uint16_t block_az_duration, double fov_blind_duration)
     : pkt_(pkt), BLOCKS_PER_PKT(blocks_per_pkt), BLOCK_DURATION(block_duration), 
     BLOCK_AZ_DURATION(block_az_duration), FOV_BLIND_DURATION(fov_blind_duration) 
   {
@@ -58,13 +59,14 @@ public:
   }
 
 protected:
+
   const T_Packet& pkt_;
   const uint16_t BLOCKS_PER_PKT;
   const double BLOCK_DURATION;
   const uint16_t BLOCK_AZ_DURATION;
-  const float FOV_BLIND_DURATION;
+  const double FOV_BLIND_DURATION;
   int32_t az_diffs[MAX_BLOCKS_PER_PKT];
-  float tss[MAX_BLOCKS_PER_PKT];
+  double tss[MAX_BLOCKS_PER_PKT];
 };
 
 template <typename T_Packet>
@@ -73,14 +75,14 @@ class SingleReturnBlockIterator : public BlockIterator<T_Packet>
 public:
 
   SingleReturnBlockIterator(const T_Packet& pkt, uint16_t blocks_per_pkt, double block_duration, 
-      uint16_t block_az_duration, float fov_blind_duration)
+      uint16_t block_az_duration, double fov_blind_duration)
     : BlockIterator<T_Packet>(pkt, blocks_per_pkt, block_duration, block_az_duration, fov_blind_duration) 
   {
-    float tss = 0;
+    double tss = 0;
     uint16_t blk = 0;
     for (; blk < (this->BLOCKS_PER_PKT - 1); blk++)
     {
-      float ts_diff = this->BLOCK_DURATION;
+      double ts_diff = this->BLOCK_DURATION;
       int32_t az_diff = ntohs(this->pkt_.blocks[blk+1].azimuth) - ntohs(this->pkt_.blocks[blk].azimuth);
       if (az_diff < 0) { az_diff += 36000; }
 
@@ -108,14 +110,14 @@ class DualReturnBlockIterator : public BlockIterator<T_Packet>
 public:
 
   DualReturnBlockIterator(const T_Packet& pkt, uint16_t blocks_per_pkt, double block_duration,
-      uint16_t block_az_duration, float fov_blind_duration)
+      uint16_t block_az_duration, double fov_blind_duration)
     : BlockIterator<T_Packet>(pkt, blocks_per_pkt, block_duration, block_az_duration, fov_blind_duration) 
   {
-    float tss = 0;
+    double tss = 0;
     uint16_t blk = 0;
     for (; blk < (this->BLOCKS_PER_PKT - 2); blk = blk + 2)
     {
-      float ts_diff = this->BLOCK_DURATION;
+      double ts_diff = this->BLOCK_DURATION;
       int32_t az_diff = ntohs(this->pkt_.blocks[blk+2].azimuth) - ntohs(this->pkt_.blocks[blk].azimuth);
       if (az_diff < 0) { az_diff += 36000; }
 
@@ -143,10 +145,10 @@ class ABDualReturnBlockIterator : public BlockIterator<T_Packet>
 public:
 
   ABDualReturnBlockIterator(const T_Packet& pkt, uint16_t blocks_per_pkt, double block_duration,
-      uint16_t block_az_duration, float fov_blind_duration)
+      uint16_t block_az_duration, double fov_blind_duration)
     : BlockIterator<T_Packet>(pkt, blocks_per_pkt, block_duration, block_az_duration, fov_blind_duration) 
   {
-    float ts_diff = this->BLOCK_DURATION;
+    double ts_diff = this->BLOCK_DURATION;
     int32_t az_diff = ntohs(this->pkt_.blocks[2].azimuth) - ntohs(this->pkt_.blocks[0].azimuth);
     if (az_diff < 0) { az_diff += 36000; }
 
@@ -159,7 +161,7 @@ public:
 
     if (ntohs(this->pkt_.blocks[0].azimuth) == ntohs(this->pkt_.blocks[1].azimuth)) // AAB
     {
-      float tss = 0;
+      double tss = 0;
       this->az_diffs[0] = this->az_diffs[1] = az_diff;
       this->tss[0] = this->tss[1] = tss;
 
@@ -169,7 +171,7 @@ public:
     }
     else // ABB
     {
-      float tss = 0;
+      double tss = 0;
       this->az_diffs[0] = az_diff;
       this->tss[0] = tss;
 
@@ -191,19 +193,19 @@ public:
     for (uint16_t chan = 0; chan < 32; chan++)
     {
       az_percents[chan] = firing_tss[chan] / (blk_ts * 2);
-      ts_diffs[chan] = firing_tss[chan] / 1000000;
+      ts_diffs[chan] = (double)firing_tss[chan] / 1000000;
     }
   }
 
   Rs16SingleReturnBlockIterator(const T_Packet& pkt, uint16_t blocks_per_pkt, double block_duration, 
-      uint16_t block_az_duration, float fov_blind_duration)
+      uint16_t block_az_duration, double fov_blind_duration)
     : BlockIterator<T_Packet>(pkt, blocks_per_pkt, block_duration, block_az_duration, fov_blind_duration) 
   {
-    float tss = 0;
+    double tss = 0;
     uint16_t blk = 0;
     for (; blk < (this->BLOCKS_PER_PKT - 1); blk++)
     {
-      float ts_diff = this->BLOCK_DURATION * 2;
+      double ts_diff = this->BLOCK_DURATION * 2;
       int32_t az_diff = ntohs(this->pkt_.blocks[blk+1].azimuth) - ntohs(this->pkt_.blocks[blk].azimuth);
       if (az_diff < 0) { az_diff += 36000; }
 
@@ -236,19 +238,19 @@ public:
     for (uint16_t chan = 0; chan < 32; chan++)
     {
       az_percents[chan] = firing_tss[chan%16] / blk_ts;
-      ts_diffs[chan] = firing_tss[chan%16] / 1000000;
+      ts_diffs[chan] = (double)firing_tss[chan%16] / 1000000;
     }
   }
 
   Rs16DualReturnBlockIterator(const T_Packet& pkt, uint16_t blocks_per_pkt, double block_duration,
-      uint16_t block_az_duration, float fov_blind_duration)
+      uint16_t block_az_duration, double fov_blind_duration)
     : BlockIterator<T_Packet>(pkt, blocks_per_pkt, block_duration, block_az_duration, fov_blind_duration) 
   {
-    float tss = 0;
+    double tss = 0;
     uint16_t blk = 0;
     for (; blk < (this->BLOCKS_PER_PKT - 1); blk++)
     {
-      float ts_diff = this->BLOCK_DURATION;
+      double ts_diff = this->BLOCK_DURATION;
       int32_t az_diff = ntohs(this->pkt_.blocks[blk+1].azimuth) - ntohs(this->pkt_.blocks[blk].azimuth);
       if (az_diff < 0) { az_diff += 36000; }
