+++ src/rs_driver/driver/decoder/decoder_RS32.hpp
@@ -95,14 +95,14 @@ inline void RS32DifopPkt2Adapter (const RS32DifopPkt& src, AdapterDifopPkt& dst)
     dst.vert_angle_cali[i].sign = src.vert_angle_cali[i].sign;
 
     v = ntohs(src.vert_angle_cali[i].value);
-    v = std::round(v * 0.1f);
+    v = (uint16_t)std::round(v * 0.1f);
     dst.vert_angle_cali[i].value = htons(v);
 
     // horiz_angles
     dst.horiz_angle_cali[i].sign = src.horiz_angle_cali[i].sign;
 
     v = ntohs(src.horiz_angle_cali[i].value);
-    v = std::round(v * 0.1f);
+    v = (uint16_t)std::round(v * 0.1f);
     dst.horiz_angle_cali[i].value = htons(v);
   }
 }
@@ -272,7 +272,7 @@ inline bool DecoderRS32<T_PointCloud>::internDecodeMsopPkt(const uint8_t* packet
     }
 
     int32_t block_az_diff;
-    float block_ts_off;
+    double block_ts_off;
     iter.get(blk, block_az_diff, block_ts_off);
 
     for (uint16_t chan = 0; chan < this->const_param_.CHANNELS_PER_BLOCK; chan++)
