+++ test/ab_dual_return_block_iterator_test.cpp
@@ -41,7 +41,7 @@ TEST(TestABDualPacketTraverser, ctor)
         2.0f); // fov_blind_duration
 
     int32_t az_diff;
-    float ts;
+    double ts;
 
     // first block
     iter.get (0, az_diff, ts);
@@ -75,7 +75,7 @@ TEST(TestABDualPacketTraverser, ctor)
         2.0f); // fov_blind_duration
 
     int32_t az_diff;
-    float ts;
+    double ts;
 
     // first block
     iter.get (0, az_diff, ts);
@@ -113,7 +113,7 @@ TEST(TestABDualPacketTraverser, ctor_fov)
         2.0f); // fov_blind_duration
 
     int32_t az_diff;
-    float ts;
+    double ts;
 
     // first block
     iter.get (0, az_diff, ts);
