+++ test/single_return_block_iterator_test.cpp
@@ -39,7 +39,7 @@ TEST(TestSingleReturnBlockIterator, ctor)
       2.0f); // fov_blind_duration
 
   int32_t az_diff;
-  float ts;
+  double ts;
 
   // first block
   iter.get (0, az_diff, ts);
@@ -71,7 +71,7 @@ TEST(TestSingleReturnBlockIterator, ctor_fov)
       2.0f); // fov_blind_duration
 
   int32_t az_diff;
-  float ts;
+  double ts;
 
   // first block
   iter.get (0, az_diff, ts);
